binning off 4-list functions and refactoring them
