+++ nexus_constructor/nexus/nexus_wrapper.py
@@ -3,11 +3,8 @@ import uuid
 
 import h5py
 from PySide2.QtCore import Signal, QObject
-from typing import Any, TypeVar, Optional, List
+from typing import Any, TypeVar, Optional
 import numpy as np
-
-from nexus_constructor.invalid_field_names import INVALID_FIELD_NAMES
-from nexus_constructor.validators import FieldType
 
 h5Node = TypeVar("h5Node", h5py.Group, h5py.Dataset)
 
@@ -53,53 +50,6 @@ def create_temporary_in_memory_file() -> h5py.File:
     :return: The file object
     """
     return set_up_in_memory_nexus_file(str(uuid.uuid4()))
-
-
-def get_fields(
-    group: h5py.Group
-) -> (List[h5py.Dataset], List[h5py.Dataset], List[h5py.Group], List[h5py.Group]):
-    """
-    Return a list of fields in a given component group.
-    :param group: The hdf5 component group to check for fields
-    :return: A list of a fields, regardless of field type
-    """
-    scalar_fields = []
-    array_fields = []
-    stream_fields = []
-    link_fields = []
-    for item in group.values():
-        item, field_type = find_field_type(item)
-        if field_type == FieldType.scalar_dataset:
-            scalar_fields.append(item)
-        elif field_type == FieldType.array_dataset:
-            array_fields.append(item)
-        elif field_type == FieldType.link:
-            link_fields.append(item)
-        elif field_type == FieldType.kafka_stream:
-            stream_fields.append(item)
-
-    return scalar_fields, array_fields, stream_fields, link_fields
-
-
-def find_field_type(item):
-    if (
-        isinstance(item, h5py.Dataset)
-        and get_name_of_node(item) not in INVALID_FIELD_NAMES
-    ):
-        if np.isscalar(item[()]):
-            return item, FieldType.scalar_dataset
-        else:
-            return item, FieldType.array_dataset
-
-    elif isinstance(item, h5py.Group):
-        if isinstance(item.parent.get(item.name, getlink=True), h5py.SoftLink):
-            return item, FieldType.link
-        elif "NX_class" in item.attrs.keys() and item.attrs["NX_class"] == "NCstream":
-            return item, FieldType.kafka_stream
-    logging.debug(
-        f"Object {get_name_of_node(item)} not handled as field - could be used for other parts of UI instead"
-    )
-    return item, None
 
 
 class NexusWrapper(QObject):
